# Conflicts:
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/AstronomerModel.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/AstronomerModelImpl.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/controller/InfoFragment.java
#	app/src/main/java/edu/cnm/deepdive/celestialbodies/model/entity/StarDetail.java

+++ app/src/main/java/edu/cnm/deepdive/celestialbodies/model/entity/StarDetail.java
@@ -292,7 +292,7 @@ public class StarDetail {
   /**
    * Returns the star's distance in parsecs
    */
-  public double getDist() {
+  public Integer getDist() {
     return dist;
   }
